chatbot/backend/requirements.txt
langgraph
langchain
langchain-openai
fastapi
uvicorn[standard]
python-dotenv
typing_extensions 